// src/app/tmc/models/DestinationAreaType.ts
export enum DestinationAreaType
{
    /// <summary>
    /// 国内
    /// </summary>
    // [Descr("国内")]
    Domestic = 1,
    /// <summary>
    /// 东亚
    /// </summary>
    // [Descr("东亚")]
    EastAsia = 2,
    /// <summary>
    /// 东南亚
    /// </summary>
    // [Descr("东南亚")]
    SoutheastAsia = 3,
    /// <summary>
    /// 西亚
    /// </summary>
    // [Descr("西亚")]
    WestAsia = 4,
    /// <summary>
    /// 中亚
    /// </summary>
    // [Descr("中亚")]
    MiddleAsia = 5,
    /// <summary>
    /// 大洋洲
    /// </summary>
    // [Descr("大洋洲")]
    Oceanica = 6,
    /// <summary>
    /// 北美
    /// </summary>
    // [Descr("北美洲")]
    NorthAmerica = 7,
    /// <summary>
    /// 南美
    /// </summary>
    // [Descr("南美洲")]
    SouthAmerica = 8,
    /// <summary>
    /// 欧洲
    /// </summary>
    // [Descr("欧洲")]
    Europe = 9,
    /// <summary>
    /// 非洲
    /// </summary>
    // [Descr("非洲")]
    Africa = 10,
    /// <summary>
    /// 港澳台
    /// </summary>
    // [Descr("港澳台")]
    HongKongMacaoTaiwan = 11
}